Collapse actor controller/state-machine duplicates into .mts re-exports
Make controller.mts and state-machine.mts the canonical sources; the
.js twins become 1-line runtime re-export shims. Add
extractMotivationGoals() so configured motivation goals flow into
runtime-decision objectives, and update the charter's motivation-flow
line to name controller.mts as the resolution site.
